Added MSVC2003 project and dependencies
For those not in the know, Visual Studio 2003 was what Pixel used to
compile Cave Story. Getting this to compile in its original environment
goes a long way to ensuring this decompilation's accuracy, since it
allows us to check if the code creates the exact same assembly.

Since the decompilation doesn't include the original DirectX/WinAPI
backends yet, compatible SDL2 and Freetype .lib files are provided.