More versatile refresh rate detection
Unfortunately this doesn't fix whatever issue Mint found.

Any other 144Hz monitor users reading this who could confirm the game
correctly limits itself to 60FPS?

Anyway, the problem here is that SDL_GetWindowDisplayMode seems to
only be suitable for when the window is fullscreen, according to the
docs at least.