Fix a memory leak
Hard to explain unless you know exactly what the code does: basically
we use a surface's userdata pointer to store a buffer that SDL2 will
not free on its own, so then when the surface is freed, we can free
it manually. But, for some reason, we were freeing the userdata
pointer of the *converted* surface, when it should have been the
original.