Add a default controller binding for pausing
This way, you're not completely screwed if you can't use a keyboard
(like you are if you run CSE2E from the terminal on a Raspberry Pi,
where for some reason SDL2 eats the keyboard input, but doesn't
forward it to the game either, forcing you to use a controller).

But yeah, ranting aside, this allows you to quit the game with just a
controller at hand, and also allows you to change the keybindings
without needing a keyboard to open the pause menu in the first place.
I just hope your controller has enough buttons for the pause button
to be mapped to one of them. Also if your controller doesn't have
any sticks you're still screwed.