Fix GLFW3 controller rebinding
For some reason, the controller would update the second time the
GetJoystickState function is ran per frame, causing input-ghosting
where binding, say, the up key, would cause the menu to scroll up.
This doesn't happen in the SDL2 backend, where the controller state
is consistent across all GetJoystickState calls per frame.

The solution to this is to just only call GetJoystickState once per
frame, and cache the result so it can be accessed multiple times.