Change Font.cpp to use CP1252 instead of UTF-8
This is accurate to the original EXE, which would default to CP1252
depending on what region Windows was set to. This is the case for
English and Spanish regions, and likely others.

This should make this branch compatible with a number of fan-made
translations. However, this also makes it so that the `PutText`
functions will no longer support UTF-8 strings.

Modders may prefer to switch to UTF-8 entirely, so I've left the old
UTF-8 parsing function in a comment.